Stream view processor: replace ad-hoc view reducers with a proper stream processor contract
- Define StreamViewProcessorContract in packages/ui using defineProcessorContract
  with consumesAllEvents: true (same pattern as JsonataReactorProcessorContract)
- State is mode-agnostic: always accumulates both raw summaries + semantic elements
- Renderer modes (raw-pretty, pretty, raw-single-json) become pure view-time filters
  — switching modes is instant, no replay needed
- apps/os2: uses runProcessorReduce incrementally via useRef for O(N) processing
  (was O(N²) before — each SSE event re-reduced all events from scratch)
- apps/os2: fast-path appendStreamEvent skips sort for in-order events
- apps/events, apps/agents: migrated to reduceStreamViewEvents() convenience helper
- Removed EventsStreamViewReducer type, old mode-specific reducers, and ~770 lines
  of duplicated reduce logic from feed-processors.ts
- All 15 stream view tests pass against the new processor contract
