+++ apps/agents/src/routes/index.tsx
@@ -10,10 +10,7 @@ import {
 } from "@iterate-com/shared/streams/types";
 import { makeFunnySlug } from "@iterate-com/shared/slug-maker";
 import { CODEMODE_CHAT_RESPONSE_SYSTEM_PROMPT } from "@iterate-com/shared/stream-processors/legacy-codemode/contract";
-import {
-  processEventsWithViewReducer,
-  rawPrettyEventsStreamViewReducer,
-} from "@iterate-com/ui/components/events/feed-processors";
+import { reduceStreamViewEvents } from "@iterate-com/ui/components/events/stream-view-processor/contract";
 import { EventsStreamComposer } from "@iterate-com/ui/components/events/stream-composer";
 import {
   EventsStreamView,
@@ -194,11 +191,7 @@ function SelectedAgentStreamView({
     refetchOnWindowFocus: true,
   });
   const viewState = useMemo(
-    () =>
-      processEventsWithViewReducer({
-        events: eventsQuery.data ?? [],
-        reducer: rawPrettyEventsStreamViewReducer,
-      }),
+    () => reduceStreamViewEvents(eventsQuery.data ?? []),
     [eventsQuery.data],
   );
   const viewerUrl = buildStreamComposerUrl({
